catalogo-habysupply/catalogo-habysupply/requirements.txt
requests
python-dotenv
mysql-connector-python
Pillow
tqdm
